+++ packages/plugin-rax-app/src/utils/setStaticConfig.js
@@ -1,6 +1,6 @@
 const path = require('path');
 const fs = require('fs-extra');
-const { STATIC_CONFIG, CUSTOM_TAB_BAR } = require('../constants');
+const { STATIC_CONFIG, CUSTOM_TAB_BAR_PATH } = require('../constants');
 
 module.exports = (api) => {
   const { setValue, context } = api;
@@ -13,17 +13,18 @@ module.exports = (api) => {
   }
 
   if (staticConfig.tabBar && staticConfig.tabBar.custom) {
-    const customTabBarPath = path.join(rootDir, 'src/CustomTabBar/index');
-    if (!checkComponentFileExists(customTabBarPath)) {
-      throw new Error(`There need custom tab bar implement in ${customTabBarPath}`);
+    let tabBarPath = path.join(rootDir, 'src/components/CustomTabBar/index');
+    if (!checkComponentFileExists(tabBarPath)) {
+      tabBarPath = path.join(rootDir, 'src/CustomTabBar/index');
+      if (!checkComponentFileExists(tabBarPath)) {
+        throw new Error('There need custom tab bar implement in src/components/CustomTabBar/index.jsx');
+      }
     }
     if (!Array.isArray(staticConfig.tabBar.list)) {
       throw new Error('There should have list field as array type to know which page need show tab bar');
     }
-    setValue(CUSTOM_TAB_BAR, true);
-    staticConfig.tabBar.source = 'CustomTabBar/index';
-  } else {
-    setValue(CUSTOM_TAB_BAR, false);
+    staticConfig.tabBar.source = tabBarPath.replace(`${rootDir}/src`, '');
+    setValue(CUSTOM_TAB_BAR_PATH, tabBarPath);
   }
   setValue(STATIC_CONFIG, staticConfig);
 };