+++ packages/plugin-rax-app/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 6.4.0
+
+- Feat: add `exportsField` polyfill
+- Chore: change custom tab bar path to `src/components/CustomTabBar/index`
+
 ## 6.3.2
 
 - Feat: update dev and build log method for miniapps